Refactor: Extract API into isolated internal package
- Moved `api.py`, `circuit_breaker.py`, `const_api.py`, `const_devices.py`, `utils_rate_limiter.py`, and `utils_sanitizer.py` into a new `violet_pool_api` module inside the custom component.
- Untangled imports so the internal API package has absolutely zero dependencies on Home Assistant constants or the custom component's codebase.
- Re-routed all imports across `custom_components` and `tests` to use relative imports to the new `violet_pool_api` package.
- This serves as the direct preparation step for publishing the API to PyPI (`Xerolux/violet_pool_api`) while ensuring the custom component remains fully functional right now.
