+++ custom_components/violet_pool_controller/violet_pool_api/circuit_breaker.py
