+++ tests/test_offline_scenarios.py
@@ -7,7 +7,7 @@ from datetime import datetime
 from homeassistant.helpers.update_coordinator import UpdateFailed
 
 from custom_components.violet_pool_controller.device import VioletPoolControllerDevice
-from custom_components.violet_pool_controller.api import VioletPoolAPIError
+from custom_components.violet_pool_controller.violet_pool_api.api import VioletPoolAPIError
 
 
 @pytest.fixture
